refactor: Making store layer only transactional
