+++ server/services/v1/service.go
@@ -33,7 +33,7 @@ type Service interface {
 	RegisterGRPC(grpc *grpc.Server) error
 }
 
-func GetRegisteredServicesRealtime(kvStore kv.KeyValueStore, searchStore search.Store, tenantMgr *metadata.TenantManager, txMgr *transaction.Manager) []Service {
+func GetRegisteredServicesRealtime(kvStore kv.TxStore, searchStore search.Store, tenantMgr *metadata.TenantManager, txMgr *transaction.Manager) []Service {
 	var v1Services []Service
 	v1Services = append(v1Services, newRealtimeService(kvStore, searchStore, tenantMgr, txMgr))
 	v1Services = append(v1Services, newHealthService(txMgr))
@@ -41,7 +41,7 @@ func GetRegisteredServicesRealtime(kvStore kv.KeyValueStore, searchStore search.
 	return v1Services
 }
 
-func GetRegisteredServices(kvStore kv.KeyValueStore, searchStore search.Store, tenantMgr *metadata.TenantManager, txMgr *transaction.Manager) []Service {
+func GetRegisteredServices(kvStore kv.TxStore, searchStore search.Store, tenantMgr *metadata.TenantManager, txMgr *transaction.Manager) []Service {
 	var v1Services []Service
 	v1Services = append(v1Services, newHealthService(txMgr))
